isolate RegexMapping interface
